Rename SearchView folder -> SearchInput
