+++ src/components/shared/searchInput/NodeItem.tsx
