+++ src/components/shared/searchInput/SearchInput.tsx
@@ -22,8 +22,7 @@ class SearchInput extends React.Component<ISearchInputProps> {
     render() {
         const isDisabled = this.props.searchStore!.isSearchDisabled;
         return (
-            // TODO: rename as searchView / searchBar
-            <div className={s.lineSearchView}>
+            <div className={s.searchInput}>
                 <div className={classnames(s.inputContainer, isDisabled ? s.disabled : undefined)}>
                     <input
                         placeholder='Hae'
