+++ src/components/shared/searchInput/nodeItem.scss
