+++ src/components/shared/searchInput/searchInput.scss
@@ -1,4 +1,4 @@
-.lineSearchView {
+.searchInput {
     .inputContainer {
         padding: 10px;
         position: relative;
