+++ src/components/shared/searchInput/searchResults.scss
