+++ src/components/sidebar/routeListView/RouteListView.tsx
@@ -29,8 +29,8 @@ import { SearchStore } from '~/stores/searchStore';
 import NavigationUtils from '~/utils/NavigationUtils';
 import TransitToggleButtonBar from '../../controls/TransitToggleButtonBar';
 import Loader from '../../shared/loader/Loader';
-import SearchInput from '../../shared/searchView/SearchInput';
-import SearchResults from '../../shared/searchView/SearchResults';
+import SearchInput from '../../shared/searchInput/SearchInput';
+import SearchResults from '../../shared/searchInput/SearchResults';
 import SidebarHeader from '../SidebarHeader';
 import EntityTypeToggles from '../homeView/EntityTypeToggles';
 import CopyRoutePathView from './CopyRoutePathView';
